Move some spawning code to layout manager and add black background.
