+++ src/layout_manager/layout_manager.rs
@@ -3,6 +3,13 @@ use std::io::Write;
 use std::os::fd::FromRawFd;
 use std::path::PathBuf;
 
+use std::os::fd::{AsFd, IntoRawFd};
+use std::io::Read;
+use nix::fcntl;
+use nix::poll;
+use nix::unistd;
+use nix::unistd::Pid;
+
 use wayland_client::Connection;
 use wayland_client::Dispatch;
 use wayland_client::Proxy;
@@ -352,7 +359,7 @@ pub fn start_layout_manager(fd: i32, layout_width: i32, layout_height: i32) {
     };
 
     let mut has_sent_data = false;
-    let mut has_ran_startup_commands = false;
+    let mut startup_commands_stage = 0;
     loop {
         if let Err(e) = event_queue.blocking_dispatch(&mut state) {
             eprintln!("Wayland connection closed {e}");
@@ -375,17 +382,33 @@ pub fn start_layout_manager(fd: i32, layout_width: i32, layout_height: i32) {
             has_sent_data = true;
         }
 
-        if !has_ran_startup_commands {
-            if let Some(ctrl) = &state.river_control {
-                if let Some(seat) = &state.seat {
-                    // Set border-width so we dont have an area arround it.
-                    ctrl.add_argument("border-width".to_string());
-                    ctrl.add_argument("0".to_string());
-                    ctrl.run_command(seat, &qh, ());
 
-                    has_ran_startup_commands = true;
+        match startup_commands_stage {
+            0 => {
+                if let Some(ctrl) = &state.river_control {
+                    if let Some(seat) = &state.seat {
+                        // Set border-width so we dont have an area arround it.
+                        ctrl.add_argument("border-width".to_string());
+                        ctrl.add_argument("0".to_string());
+                        ctrl.run_command(seat, &qh, ());
+
+                        startup_commands_stage = 1;
+                    }
                 }
             }
+            1 => {
+                if let Some(ctrl) = &state.river_control {
+                    if let Some(seat) = &state.seat {
+                        // Set background color to more "natural" one than default light blue for games.
+                        ctrl.add_argument("background-color".to_string());
+                        ctrl.add_argument("0x000000".to_string());
+                        ctrl.run_command(seat, &qh, ());
+
+                        startup_commands_stage = 2;
+                    }
+                }
+            }
+            _ => {}
         }
     }
 }
@@ -430,4 +453,135 @@ pub fn kwin_dbus_unload_script() -> Result<(), Box<dyn Error>> {
 
     println!("[partydeck] util::kwin_dbus_unload_script - Script unloaded.");
     Ok(())
+}
+
+
+pub fn spawn_comp_and_get_display(
+    comp_executable: &str,
+    primary_monitor: Monitor,
+) -> Option<(String, String, Monitor, Pid)> {
+    let base_program_buf = env::current_exe().expect("Failed to get partydeck executable");
+    let base_program = base_program_buf.to_str()?;
+
+    let (read_fd, write_fd) = unistd::pipe().unwrap();
+
+    let flags = fcntl::FdFlag::from_bits_truncate(
+        fcntl::fcntl(&write_fd, fcntl::FcntlArg::F_GETFD).unwrap(),
+    );
+    fcntl::fcntl(
+        &write_fd,
+        fcntl::FcntlArg::F_SETFD(flags & !fcntl::FdFlag::FD_CLOEXEC),
+    )
+    .expect("Failed to open pipe to river");
+
+    let mut cmd = std::process::Command::new(comp_executable);
+
+    match comp_executable {
+        "river" => {
+            cmd.args([
+                "-c",
+                format!(
+                    "{} --internal-layout {}:{}:{}",
+                    base_program,
+                    &write_fd.into_raw_fd(),
+                    primary_monitor.width(),
+                    primary_monitor.height()
+                )
+                .as_str(),
+            ]);
+        }
+        "kwin_wayland" => {
+            cmd.args([
+                "--xwayland",
+                "--exit-with-session",
+                "--height",
+                &primary_monitor.height().to_string(),
+                "--width",
+                &primary_monitor.width().to_string(),
+                "--",
+                format!(
+                    "{} --internal-layout {}:{}:{}",
+                    base_program,
+                    &write_fd.into_raw_fd(),
+                    primary_monitor.width(),
+                    primary_monitor.height()
+                )
+                .as_str(),
+            ]);
+        }
+        _ => {
+            println!(
+                "Unknown comp ({}); trusting that it works, MAY FAIL",
+                comp_executable
+            );
+            cmd.args([
+                "--",
+                format!(
+                    "{} --internal-layout {}:{}:{}",
+                    base_program,
+                    &write_fd.into_raw_fd(),
+                    primary_monitor.width(),
+                    primary_monitor.height()
+                )
+                .as_str(),
+            ]);
+        }
+    }
+
+    let child_pid;
+    match cmd.spawn() {
+        Ok(child) => child_pid = Pid::from_raw((child.id()) as i32),
+        Err(e) => {
+            eprintln!(
+                "[partydeck] Failed to start COMP ({}): {}",
+                comp_executable, e
+            );
+            return None;
+        }
+    }
+
+    let mut fds = [poll::PollFd::new(read_fd.as_fd(), poll::PollFlags::POLLIN)];
+    let res = poll::poll(&mut fds, 2000 as u16).unwrap_or(0);
+    if res == 0 {
+        eprintln!("[partydeck] NO DATA FROM COMP HANDLER");
+        return None;
+    }
+
+    let mut reader = unsafe { std::fs::File::from_raw_fd(read_fd.into_raw_fd()) };
+
+    let mut len_buf = [0u8; 4];
+    reader.read_exact(&mut len_buf).expect("Failed to read FD");
+    let way_disp_len = u32::from_be_bytes(len_buf) as usize;
+    reader.read_exact(&mut len_buf).expect("Failed to read FD");
+    let x11_disp_len = u32::from_be_bytes(len_buf) as usize;
+
+    reader.read_exact(&mut len_buf).expect("Failed to read FD");
+    let monitor_width = u32::from_be_bytes(len_buf);
+    reader.read_exact(&mut len_buf).expect("Failed to read FD");
+    let monitor_height = u32::from_be_bytes(len_buf);
+    let remote_monitor = Monitor::new("REMOTE_MONITOR".to_owned(), monitor_width, monitor_height);
+
+    let mut way_disp_buf = vec![0u8; way_disp_len];
+    reader
+        .read_exact(&mut way_disp_buf)
+        .expect("Failed to read FD");
+    let mut x11_disp_buf = vec![0u8; x11_disp_len];
+    reader
+        .read_exact(&mut x11_disp_buf)
+        .expect("Failed to read FD");
+
+    let way_disp = String::from_utf8(way_disp_buf).expect("Failed to decode FD");
+    let x11_disp = String::from_utf8(x11_disp_buf).expect("Failed to decode FD");
+
+    println!(
+        "Got DISPLAY_WAYLAND from COMP: {} and DISPLAY: {}, with resolution: {}x{}",
+        way_disp, x11_disp, monitor_width, monitor_height
+    );
+
+    return Some((
+        way_disp.to_string(),
+        x11_disp.to_string(),
+        remote_monitor,
+        child_pid,
+    ));
 }
